+++ database/migrations/positions/0002_00_01_000001_create_position_info_table.php
@@ -14,7 +14,7 @@ class CreatePositionInfoTable extends Migration
     public function up(): void
     {
         Schema::create('position_info', static function (Blueprint $table) {
-            $table->unsignedBigInteger('position_id')->unique()->primary();
+            $table->unsignedSmallInteger('position_id')->unique()->primary();
 
             $table->string('email');
             $table->string('work_phone')->nullable();
